free: add free_tiny
It is now possible to free tiny allocations.

+++ srcs/malloc_tiny.h
@@ -6,7 +6,7 @@
 /*   By: rotrojan <rotrojan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/16 19:52:54 by rotrojan          #+#    #+#             */
-/*   Updated: 2026/04/21 16:04:06 by rotrojan         ###   ########.fr       */
+/*   Updated: 2026/04/22 10:13:29 by rotrojan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@
  * We want to hold up to a hundred of 128 bytes TINY chunks. We can store 128
  * of these in a 4 pages zone.
  */
-#define TINY_ZONE_SIZE (sysconf(_SC_PAGESIZE) * 4)
+#define TINY_ZONE_SIZE (size_t)(sysconf(_SC_PAGESIZE) * 4)
 
 typedef struct tiny_zone {
 	s_zone_hdr        zone_hdr;
